feat(datastore): Migrate to Amplify Swift V2 (#4962)

+++ packages/amplify_datastore/ios/Classes/types/hub/FlutterHubElement.swift
@@ -4,15 +4,12 @@
 import Foundation
 import Flutter
 import UIKit
-import Amplify
-import AmplifyPlugins
-import AWSCore
 import Combine
 
 public struct FlutterHubElement {
     var model: [String: Any]
     var version: Int?
-    var lastChangedAt: Int?
+    var lastChangedAt: Int64?
     var deleted: Bool
 
     init(
@@ -52,11 +49,11 @@ public struct FlutterHubElement {
             self.version = hubElement.version
             self.deleted = self.model["_deleted"] as? Bool ?? false
             if let value = self.model["_lastChangedAt"] as? Double {
-                self.lastChangedAt = Int(value)
+                self.lastChangedAt = Int64(value)
             } else if let value = self.model["_lastChangedAt"] as? String {
-                self.lastChangedAt = Int(value)
+                self.lastChangedAt = Int64(value)
             } else if let value = self.model["_lastChangedAt"] as? Int {
-                self.lastChangedAt = value
+                self.lastChangedAt = Int64(value)
             }
         } catch {
             throw FlutterDataStoreError.hubEventCast
